SentryPoint_Backend/requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
supabase==2.28.3
python-dotenv==1.2.2
python-multipart==0.0.26
pydantic==2.13.3
httpx==0.28.1
requests==2.33.1
# PostGIS and geometry support
postgrest==2.28.3
Pillow==11.0.0
